// FinalProjectFrontEnd/types/theme.ts
import { MD3Colors } from 'react-native-paper/lib/typescript/types';

export interface ThemeColors extends MD3Colors {
  cardColor?: string;
  CardColorText?: string;
  buttonColor?: string;
  iconColor?: string;
  buttonTextColor?: string;
  dropDownSelected?: string;
  modalHandle?: string;
  modalColor?: string;
  modalText?: string;
  rangeTextColor?: string;
  selectedColor?: string;
  warning?: string;
  success?: string;
}